Moving version to a cleaner spot.
